# PIM-7532: improve the standard format for products associated to avoid performance impact