--- a/CHANGELOG-1.7.md
+++ b/CHANGELOG-1.7.md
@@ -2,6 +2,7 @@
 
 ## Bug fixes
 
+- PIM-7532: Improve the standard format for products associated to avoid performance impact
 - PIM-7540: Fix translations of boolean attributes
 
 # 1.7.27 (2018-07-24)