--- a/src/Pim/Bundle/CatalogBundle/Resources/config/serializers_standard.yml
+++ b/src/Pim/Bundle/CatalogBundle/Resources/config/serializers_standard.yml
@@ -42,6 +42,8 @@ services:
 
     pim_catalog.normalizer.standard.product.associations:
         class: '%pim_catalog.normalizer.standard.product.associations.class%'
+        arguments:
+            - '@pim_catalog.query.get_associated_product_codes_by_product'
         tags:
             - { name: pim_serializer.normalizer, priority: 40 }