--- a/src/Pim/Bundle/CatalogBundle/Resources/config/storage_driver/doctrine/orm.yml
+++ b/src/Pim/Bundle/CatalogBundle/Resources/config/storage_driver/doctrine/orm.yml
@@ -35,6 +35,8 @@ parameters:
     pim_catalog.doctrine.query.sorter.in_group.class:      Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\InGroupSorter
     pim_catalog.doctrine.query.sorter.is_associated.class: Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\IsAssociatedSorter
 
+    pim_catalog.query.get_associated_product_codes_by_product.class: Pim\Bundle\CatalogBundle\Doctrine\ORM\Query\GetAssociatedProductCodesByProductFromDB
+
 services:
     pim_catalog.object_manager.product:
         alias: doctrine.orm.entity_manager
@@ -140,3 +142,9 @@ services:
                 - 'LOWER THAN ON ALL LOCALES'
         tags:
             - { name: 'pim_catalog.doctrine.query.filter', priority: 30 }
+
+    pim_catalog.query.get_associated_product_codes_by_product:
+        class: '%pim_catalog.query.get_associated_product_codes_by_product.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.association.class%'