--- a/src/Pim/Bundle/CatalogBundle/tests/integration/Doctrine/ORM/Query/GetAssociatedProductCodesByProductIntegration.php
+++ b/src/Pim/Bundle/CatalogBundle/tests/integration/Doctrine/ORM/Query/GetAssociatedProductCodesByProductIntegration.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace Pim\Bundle\CatalogBundle\tests\integration\Doctrine\ORM\Query;
+
+use Akeneo\Test\Integration\Configuration;
+use Akeneo\Test\Integration\TestCase;
+
+class GetAssociatedProductCodesByProductIntegration extends TestCase
+{
+    public function testQueryToGetAssociatedProductCodes()
+    {
+        $productA = $this->get('pim_catalog.builder.product')->createProduct('productA');
+        $productB = $this->get('pim_catalog.builder.product')->createProduct('productB');
+        $productC = $this->get('pim_catalog.builder.product')->createProduct('productC');
+        $productD = $this->get('pim_catalog.builder.product')->createProduct('productD');
+
+        $this->get('pim_catalog.saver.product')->saveAll([$productB, $productC, $productD]);
+
+        $this->get('pim_catalog.updater.product')->update($productA, [
+            'associations' => [
+                'X_SELL' => ['products' => ['productB']],
+                'PACK' => ['products' => ['productC', 'productD']],
+                'UPSELL' => ['products' => []],
+            ]
+        ]);
+        $this->get('pim_catalog.saver.product')->save($productA);
+
+        $associationTypeRepository = $this->get('pim_catalog.repository.association_type');
+        $xsell = $associationTypeRepository->findOneByIdentifier('X_SELL');
+        $pack = $associationTypeRepository->findOneByIdentifier('PACK');
+        $upsell = $associationTypeRepository->findOneByIdentifier('UPSELL');
+
+        $query = $this->get('pim_catalog.query.get_associated_product_codes_by_product');
+        $this->assertSame(['productB'], $query->getCodes(4, $xsell->getId()));
+        $this->assertSame(['productC', 'productD'], $query->getCodes(4, $pack->getId()));
+        $this->assertSame([], $query->getCodes(4, $upsell->getId()));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getConfiguration()
+    {
+        return new Configuration([Configuration::getTechnicalCatalogPath()]);
+    }
+}